Changed variable to correct type.
Uniform location variables are declared as GLuint but "glGetUniformLocation" returns a GLint and the first parameter of "glUniform1f" requires a GLint. I changed the variables to type GLint in both the cpp files and in the example code in the docs. This helps knock out the sign-conversion warnings which I've told my compiler to treat as errors.